backend/requirements.txt
# FastAPI server (with Pydantic v2 support)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Data validation (Pydantic v2)
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MongoDB async driver
motor>=3.1.0
pymongo>=4.0.0

# HTTP client
httpx>=0.24.0

# AI / OpenAI APIs
openai>=1.8.0
tenacity>=8.2.0

# Environment variables
python-dotenv>=1.0.0

# Security
passlib[bcrypt]>=1.7.0
python-jose>=3.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0